requirements.txt
Flask
flask-socketio
Werkzeug
Jinja2
python-dotenv
sentence-transformers
scikit-learn
groq
numpy
torch
flask-sqlalchemy
flask-migrate
sqlalchemy[asyncio]
asyncpg
psycopg[binary,pool]
Authlib>=1.3.0